fix: suppress double-promotion warnings in nsec2timespec callers
The nsec2timespec parameter was widened to long double to avoid
precision loss on the Apple path.  Callers that pass double values now
need explicit (long double) casts (limit_process.c) or L-suffixed
literals (cpulimit_test.c) to suppress the -Wdouble-promotion
diagnostic that clang-tidy raises for implicit float→long double
promotions.

Also improve the malloc failure message in multi_process_busy.c.

Agent-Logs-Url: https://github.com/HiGarfield/cpulimit/sessions/4e06bdc4-d942-4fd1-8a0a-d57701a1759c

+++ src/limit_process.c
@@ -313,10 +313,10 @@ void limit_process(pid_t pid, double limit, int include_children, int verbose) {
 
         /* Split time slot into work and sleep periods */
         work_time_ns = time_slot * 1000 * work_ratio;
-        nsec2timespec(work_time_ns, &work_time);
+        nsec2timespec((long double)work_time_ns, &work_time);
 
         sleep_time_ns = time_slot * 1000 - work_time_ns;
-        nsec2timespec(sleep_time_ns, &sleep_time);
+        nsec2timespec((long double)sleep_time_ns, &sleep_time);
 
         if (verbose) {
             if (cycle_counter % STATS_SAMPLE_PERIOD == 0) {
